requirements.txt
fastapi==0.104.1
pandas==2.0.2
pydantic==2.5.1
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.3
uvicorn==0.24.0.post1
requests==2.31.0
